Changed UI of login page
