--- client/expense-manager/src/AuthForm.css
@@ -1,335 +1,0 @@
-*,
-*:before,
-*:after {
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-}
-
-/* body {
-    font-family: 'Open Sans', Helvetica, Arial, sans-serif;
-    background: #ffffff;
-} */
-
-.inputl,
-.buttonl {
-    border: none;
-    outline: none;
-    background: none;
-    /* font-family: 'Open Sans', Helvetica, Arial, sans-serif; */
-}
-
-.tip {
-    font-size: 20px;
-    margin: 40px auto 50px;
-    text-align: center;
-}
-
-.cont {
-    border-radius: 20px;
-    overflow: hidden;
-    position: relative;
-    width: 900px;
-    height: 550px;
-    margin: 0 auto 100px;
-    background: transparent;
-    box-shadow: 0 0 50px #0ef
-}
-
-.form {
-    position: relative;
-    width: 640px;
-    height: 100%;
-    -webkit-transition: -webkit-transform 1.2s ease-in-out;
-    transition: -webkit-transform 1.2s ease-in-out;
-    transition: transform 1.2s ease-in-out;
-    transition: transform 1.2s ease-in-out, -webkit-transform 1.2s ease-in-out;
-    /* padding: 50px 30px 0; */
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-evenly;
-}
-
-.sub-cont {
-    overflow: hidden;
-    position: absolute;
-    left: 640px;
-    top: 0;
-    width: 900px;
-    height: 100%;
-    padding-left: 260px;
-    background: transparent;
-    -webkit-transition: -webkit-transform 1.2s ease-in-out;
-    transition: -webkit-transform 1.2s ease-in-out;
-    transition: transform 1.2s ease-in-out;
-    transition: transform 1.2s ease-in-out, -webkit-transform 1.2s ease-in-out;
-}
-
-.cont.s--signup .sub-cont {
-    -webkit-transform: translate3d(-640px, 0, 0);
-    transform: translate3d(-640px, 0, 0);
-}
-
-.buttonl {
-    display: block;
-    margin: 0 auto;
-    width: 260px;
-    height: 36px;
-    border-radius: 15px;
-    color: black;
-    border: 1px solid #0ef;
-    background-color: #0ef;
-    font-size: 15px;
-    cursor: pointer;
-}
-
-.buttonll{
-    font-size: 15px;
-    color: rgb(51, 186, 240);
-    background-color: transparent;
-    border: none;
-    outline: none;
-    text-decoration: underline;
-}
-
-.img {
-    overflow: hidden;
-    z-index: 2;
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 260px;
-    height: 100%;
-    padding-top: 180px;
-}
-
-.img:before {
-    content: '';
-    position: absolute;
-    right: 0;
-    top: 0;
-    width: 900px;
-    height: 100%;
-    /* background-image: url("ext.jpg"); */
-    opacity: .8;
-    background-size: cover;
-    -webkit-transition: -webkit-transform 1.2s ease-in-out;
-    transition: transform 1.2s ease-in-out;
-    transition: transform 1.2s ease-in-out, -webkit-transform 1.2s ease-in-out;
-}
-
-.img:after {
-    content: '';
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.6);
-}
-
-.cont.s--signup .img:before {
-    -webkit-transform: translate3d(640px, 0, 0);
-    transform: translate3d(640px, 0, 0);
-}
-
-.img__text {
-    z-index: 2;
-    position: absolute;
-    left: 0;
-    top: 50px;
-    width: 100%;
-    padding: 0 20px;
-    text-align: center;
-    color: #fff;
-    -webkit-transition: -webkit-transform 1.2s ease-in-out;
-    transition: -webkit-transform 1.2s ease-in-out;
-    transition: transform 1.2s ease-in-out;
-    transition: transform 1.2s ease-in-out, -webkit-transform 1.2s ease-in-out;
-}
-
-.img__text h2 {
-    margin-bottom: 10px;
-    font-weight: normal;
-}
-
-.img__text p {
-    font-size: 14px;
-    line-height: 1.5;
-}
-
-.cont.s--signup .img__text.m--up {
-    -webkit-transform: translateX(520px);
-    transform: translateX(520px);
-}
-
-.img__text.m--in {
-    -webkit-transform: translateX(-520px);
-    transform: translateX(-520px);
-}
-
-.cont.s--signup .img__text.m--in {
-    -webkit-transform: translateX(0);
-    transform: translateX(0);
-}
-
-.img__btn {
-    overflow: hidden;
-    z-index: 2;
-    position: relative;
-    width: 100px;
-    height: 36px;
-    margin: 0 auto;
-    background: transparent;
-    color: #fff;
-    text-transform: uppercase;
-    font-size: 15px;
-    cursor: pointer;
-}
-
-.img__btn:after {
-    content: '';
-    z-index: 2;
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    border: 2px solid #fff;
-    border-radius: 30px;
-}
-
-.img__btn span {
-    position: absolute;
-    left: 0;
-    top: 0;
-    display: -webkit-box;
-    display: flex;
-    -webkit-box-pack: center;
-    justify-content: center;
-    -webkit-box-align: center;
-    align-items: center;
-    width: 100%;
-    height: 100%;
-    -webkit-transition: -webkit-transform 1.2s;
-    transition: -webkit-transform 1.2s;
-    transition: transform 1.2s;
-    transition: transform 1.2s, -webkit-transform 1.2s;
-}
-
-.img__btn span.m--in {
-    -webkit-transform: translateY(-72px);
-    transform: translateY(-72px);
-}
-
-.cont.s--signup .img__btn span.m--in {
-    -webkit-transform: translateY(0);
-    transform: translateY(0);
-}
-
-.cont.s--signup .img__btn span.m--up {
-    -webkit-transform: translateY(72px);
-    transform: translateY(72px);
-}
-
-.headingl {
-    color: #0ef;
-    width: 100%;
-    font-size: 26px;
-    text-align: center;
-}
-
-.labell {
-    /* display: block; */
-    background-color: transparent;
-    color: white;
-    margin-bottom: 15px;
-    position: relative;
-    border-bottom: 1px solid #0ef;
-    /* text-align: center; */
-}
-
-.labell span {
-
-    position:absolute;
-    top:0;
-    left:0;
-    transform:translateY(40px);
-    font-size:16px;
-    transition-duration:300ms;
-}
-
-.inputl {
-    color: #0ef;
-    width:300px;
-    padding:10px 0px;
-    margin-top:20px;
-    border:none;
-    outline:none;
-    appearance: none;
-    background-color: transparent;
-    font-size: 16px;
-}
-
-.inputl::placeholder{
-    opacity: 0;
-}
-
-.labell:focus-within > span,
-.inputl:not(:placeholder-shown) + span{
-    color:white;
-    transform:translateY(0px);
-}
-
-.forgot-pass {
-    margin-top: 15px;
-    text-align: center;
-    font-size: 12px;
-    color: black;
-}
-
-.submit {
-    margin-top: 40px;
-    margin-bottom: 20px;
-    background: #d4af7a;
-    text-transform: uppercase;
-}
-
-.fb-btn {
-    border: 2px solid #d3dae9;
-    color: #8fa1c7;
-}
-
-.fb-btn span {
-    font-weight: bold;
-    color: #455a81;
-}
-
-.sign-in {
-    -webkit-transition-timing-function: ease-out;
-    transition-timing-function: ease-out;
-}
-
-.cont.s--signup .sign-in {
-    -webkit-transition-timing-function: ease-in-out;
-    transition-timing-function: ease-in-out;
-    -webkit-transition-duration: 1.2s;
-    transition-duration: 1.2s;
-    -webkit-transform: translate3d(890px, 0, 0);
-    transform: translate3d(890px, 0, 0);
-}
-
-.sign-up {
-    -webkit-transform: translate3d(-900px, 0, 0);
-    transform: translate3d(-900px, 0, 0);
-}
-
-.cont.s--signup .sign-up {
-    -webkit-transform: translate3d(0, 0, 0);
-    transform: translate3d(0, 0, 0);
-}
-#google, #googlee{
-    display: flex;
-    justify-content: center ;
-}
